# backend/src/routes/attendance.js
const express = require('express');
const router = express.Router();
const { requireAuth } = require('../middleware/auth');

// Helper: Get UTC start and end of a local date (YYYY-MM format)
function getMonthRange(yearMonth) {
  const [year, month] = yearMonth.split('-').map(Number);
  // Convert local date to UTC range for database query
  const startDate = new Date(Date.UTC(year, month - 1, 1, 0, 0, 0, 0));
  const endDate = new Date(Date.UTC(year, month, 0, 23, 59, 59, 999));
  return { startDate, endDate };
}

// Helper: Parse YYYY-MM-DD string to UTC midnight for storage
function parseLocalDate(dateStr) {
  const [year, month, day] = dateStr.split('-').map(Number);
  return new Date(Date.UTC(year, month - 1, day));
}

// Helper: Get date string YYYY-MM-DD from Date object (local timezone for display)
function toLocalDateString(date) {
  const d = new Date(date);
  const year = d.getFullYear();
  const month = String(d.getMonth() + 1).padStart(2, '0');
  const day = String(d.getDate()).padStart(2, '0');
  return `${year}-${month}-${day}`;
}

// Helper: Get today's UTC date for attendance record (uses server time)
function getTodayUTCDate() {
  const now = new Date();
  return new Date(Date.UTC(now.getUTCFullYear(), now.getUTCMonth(), now.getUTCDate()));
}

// All routes require auth
router.use(requireAuth);

// GET /api/attendance?month=YYYY-MM
router.get('/', async (req, res) => {
  const prisma = req.app.get('prisma');
  const { month } = req.query;

  if (!month || !/^\d{4}-\d{2}$/.test(month)) {
    return res.status(400).json({ error: 'month query param required (YYYY-MM)' });
  }

  try {
    const { startDate, endDate } = getMonthRange(month);

    const records = await prisma.attendance.findMany({
      where: {
        employeeId: req.employeeId,
        date: {
          gte: startDate,
          lte: endDate,
        },
      },
      include: { sessions: { orderBy: { checkIn: 'asc' } } },
      orderBy: { date: 'asc' },
    });

    // Transform dates to local timezone strings for frontend
    const transformed = records.map(r => ({
      ...r,
      date: toLocalDateString(r.date),
      sessions: r.sessions.map(s => ({
        ...s,
        checkIn: s.checkIn ? s.checkIn.toISOString() : null,
        checkOut: s.checkOut ? s.checkOut.toISOString() : null,
      })),
    }));

    res.json(transformed);
  } catch (err) {
    console.error('Attendance list error:', err);
    res.status(500).json({ error: 'Failed to fetch attendance' });
  }
});

// GET /api/attendance/today?localDate=YYYY-MM-DD
router.get('/today', async (req, res) => {
  const prisma = req.app.get('prisma');
  const { localDate } = req.query;

  // Use frontend-provided local date if available, otherwise fall back to server UTC date
  let attendanceDate;
  if (localDate && /^\d{4}-\d{2}-\d{2}$/.test(localDate)) {
    const [year, month, day] = localDate.split('-').map(Number);
    attendanceDate = new Date(Date.UTC(year, month - 1, day));
  } else {
    attendanceDate = getTodayUTCDate();
  }

  try {
    const record = await prisma.attendance.findFirst({
      where: {
        employeeId: req.employeeId,
        date: attendanceDate,
      },
      include: { sessions: { orderBy: { checkIn: 'asc' } } },
    });

    if (!record) {
      return res.json(null);
    }

    // Transform dates to local timezone strings
    const transformed = {
      ...record,
      date: toLocalDateString(record.date),
      sessions: record.sessions.map(s => ({
        ...s,
        checkIn: s.checkIn ? s.checkIn.toISOString() : null,
        checkOut: s.checkOut ? s.checkOut.toISOString() : null,
      })),
    };

    res.json(transformed);
  } catch (err) {
    console.error('Today attendance error:', err);
    res.status(500).json({ error: 'Failed to fetch today attendance' });
  }
});

// POST /api/attendance/check-in
router.post('/check-in', async (req, res) => {
  const prisma = req.app.get('prisma');
  const { latitude, longitude, address, localDateTime } = req.body;

  if (latitude === undefined || latitude === null || longitude === undefined || longitude === null) {
    return res.status(400).json({ error: 'latitude and longitude are required' });
  }

  const lat = Number(latitude);
  const lng = Number(longitude);
  if (!Number.isFinite(lat) || !Number.isFinite(lng)) {
    return res.status(400).json({ error: 'Coordinates must be numeric' });
  }
  if (Math.abs(lat) > 90 || Math.abs(lng) > 180) {
    return res.status(400).json({ error: 'Coordinates out of range' });
  }
  // Reject 0,0 coordinates (null island — invalid location)
  if (lat === 0 && lng === 0) {
    return res.status(400).json({ error: 'Invalid location coordinates' });
  }

  // Cap client-supplied timestamp drift at ±15 minutes to prevent back/future-dating
  if (localDateTime) {
    const ts = new Date(localDateTime);
    if (!Number.isNaN(ts.getTime())) {
      const driftMs = Math.abs(ts.getTime() - Date.now());
      if (driftMs > 15 * 60 * 1000) {
        return res.status(400).json({ error: 'Check-in timestamp drift too large (max 15 minutes)' });
      }
    }
  }

  // Use employee's local datetime (sent from frontend) to derive:
  // 1. Attendance record date (local date)
  // 2. checkIn timestamp (local time converted to UTC for storage)
  let attendanceDate;
  let checkInTime;

  if (localDateTime) {
    // Parse the local datetime and convert to UTC for storage
    const localDate = new Date(localDateTime);
    checkInTime = localDate; // Store local time as-is, but in UTC context

    // Extract date components for attendance record (local date)
    const year = localDate.getFullYear();
    const month = localDate.getMonth();
    const day = localDate.getDate();
    attendanceDate = new Date(Date.UTC(year, month, day));
  } else {
    // Fallback: use current UTC
    const now = new Date();
    attendanceDate = new Date(Date.UTC(now.getUTCFullYear(), now.getUTCMonth(), now.getUTCDate()));
    checkInTime = now;
  }

  try {
    // Find or create attendance record for today
    let attendance = await prisma.attendance.findFirst({
      where: { employeeId: req.employeeId, date: attendanceDate },
    });

    if (!attendance) {
      attendance = await prisma.attendance.create({
        data: {
          employeeId: req.employeeId,
          date: attendanceDate,
          status: 'Present',
        },
      });
    }

    // Create new session with employee's local time
    const session = await prisma.attendanceSession.create({
      data: {
        attendanceId: attendance.id,
        checkIn: checkInTime,
        checkInLat: latitude,
        checkInLng: longitude,
        checkInAddr: address || null,
      },
    });

    // Return full attendance with all sessions
    const fullAttendance = await prisma.attendance.findUnique({
      where: { id: attendance.id },
      include: { sessions: { orderBy: { checkIn: 'asc' } } },
    });

    // Transform for frontend
    const transformed = {
      ...fullAttendance,
      date: toLocalDateString(fullAttendance.date),
      sessions: fullAttendance.sessions.map(s => ({
        ...s,
        checkIn: s.checkIn ? s.checkIn.toISOString() : null,
        checkOut: s.checkOut ? s.checkOut.toISOString() : null,
      })),
    };

    res.status(201).json(transformed);
  } catch (err) {
    console.error('Check-in error:', err);
    res.status(500).json({ error: 'Check-in failed' });
  }
});

// PUT /api/attendance/check-out/:sessionId
router.put('/check-out/:sessionId', async (req, res) => {
  const prisma = req.app.get('prisma');
  const { sessionId } = req.params;
  const { latitude, longitude, address } = req.body;

  if (latitude === undefined || latitude === null || longitude === undefined || longitude === null) {
    return res.status(400).json({ error: 'latitude and longitude are required' });
  }

  const lat = Number(latitude);
  const lng = Number(longitude);
  if (!Number.isFinite(lat) || !Number.isFinite(lng)) {
    return res.status(400).json({ error: 'Coordinates must be numeric' });
  }
  if (Math.abs(lat) > 90 || Math.abs(lng) > 180) {
    return res.status(400).json({ error: 'Coordinates out of range' });
  }
  if (lat === 0 && lng === 0) {
    return res.status(400).json({ error: 'Invalid location coordinates' });
  }

  try {
    const session = await prisma.attendanceSession.findUnique({
      where: { id: sessionId },
      include: { attendance: true },
    });

    if (!session) {
      return res.status(404).json({ error: 'Session not found' });
    }

    if (session.attendance.employeeId !== req.employeeId) {
      return res.status(403).json({ error: 'Not authorized' });
    }

    if (session.checkOut) {
      return res.status(409).json({ error: 'Already checked out' });
    }

    const updated = await prisma.attendanceSession.update({
      where: { id: sessionId },
      data: {
        checkOut: new Date(),
        checkOutLat: latitude,
        checkOutLng: longitude,
        checkOutAddr: address || null,
      },
      include: { attendance: { include: { sessions: { orderBy: { checkIn: 'asc' } } } } },
    });

    // Transform for frontend
    const transformed = {
      ...updated,
      checkIn: updated.checkIn ? updated.checkIn.toISOString() : null,
      checkOut: updated.checkOut ? updated.checkOut.toISOString() : null,
      attendance: {
        ...updated.attendance,
        date: toLocalDateString(updated.attendance.date),
        sessions: updated.attendance.sessions.map(s => ({
          ...s,
          checkIn: s.checkIn ? s.checkIn.toISOString() : null,
          checkOut: s.checkOut ? s.checkOut.toISOString() : null,
        })),
      },
    };

    res.json(transformed);
  } catch (err) {
    console.error('Check-out error:', err);
    res.status(500).json({ error: 'Check-out failed' });
  }
});

// GET /api/attendance/all - Admin only: get all employees attendance
router.get('/all', async (req, res) => {
  const prisma = req.app.get('prisma');
  const { month } = req.query;

  // Check if admin
  const employee = await prisma.employee.findUnique({ where: { id: req.employeeId } });
  if (!employee || !employee.isAdmin) {
    return res.status(403).json({ error: 'Admin access required' });
  }

  if (!month || !/^\d{4}-\d{2}$/.test(month)) {
    return res.status(400).json({ error: 'month query param required (YYYY-MM)' });
  }

  try {
    const { startDate, endDate } = getMonthRange(month);

    const records = await prisma.attendance.findMany({
      where: {
        date: { gte: startDate, lte: endDate },
      },
      include: {
        employee: { select: { id: true, name: true, email: true, department: true } },
        sessions: { orderBy: { checkIn: 'asc' } },
      },
      orderBy: [{ date: 'asc' }, { employee: { name: 'asc' } }],
    });

    // Transform dates for frontend
    const transformed = records.map(r => ({
      ...r,
      date: toLocalDateString(r.date),
      sessions: r.sessions.map(s => ({
        ...s,
        checkIn: s.checkIn ? s.checkIn.toISOString() : null,
        checkOut: s.checkOut ? s.checkOut.toISOString() : null,
      })),
    }));

    res.json(transformed);
  } catch (err) {
    console.error('Admin attendance error:', err);
    res.status(500).json({ error: 'Failed to fetch attendance' });
  }
});

module.exports = router;